+++ e2e/cypress/support/commands.js
@@ -358,7 +358,9 @@ Cypress.Commands.add('addVideoFromURI', () => {
 Cypress.Commands.add('addHtml', () => {
   cy.get(`[data-hook*=${HTML_PLUGIN.STATIC_TOOLBAR_BUTTON}][tabindex!=-1]`).click();
   cy.get(`[data-hook*=${PLUGIN_TOOLBAR_BUTTONS.EDIT}]`).click();
-  cy.get(`[data-hook*=${HTML_PLUGIN.INPUT}]`).type(
+  cy.get(
+    `[data-hook*=${HTML_PLUGIN.INPUT}]`
+  ).type(
     '<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">The updates, insights and stories of the engineering challenges we encounter, and our way of solving them. Subscribe to our fresh, monthly newsletter and get these goodies right to your e-mail:<a href="https://t.co/0ziRSJJAxK">https://t.co/0ziRSJJAxK</a> <a href="https://t.co/nTHlsG5z2a">pic.twitter.com/nTHlsG5z2a</a></p>&mdash; Wix Engineering (@WixEng) <a href="https://twitter.com/WixEng/status/1076810144774868992?ref_src=twsrc%5Etfw">December 23, 2018</a></blockquote> <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>',
     { delay: 0 }
   );
@@ -382,6 +384,17 @@ Cypress.Commands.add('dragAndDrop', (src, dest, elem = 0) => {
         .trigger('mousemove', { which: 1, pageX: dest.x + 50, pageY: dest.y + 20 })
         .trigger('mouseup', { force: true });
     });
+});
+
+Cypress.Commands.add('dragAndDropPlugin', (src, dest) => {
+  const dataTransfer = new DataTransfer();
+  cy.get(src)
+    .trigger('mousedown')
+    .trigger('dragstart', { dataTransfer })
+    .get(dest)
+    .trigger('dragenter', { dataTransfer })
+    .trigger('dragover', { dataTransfer })
+    .trigger('drop', { dataTransfer });
 });
 
 // disable screenshots in debug mode. So there is no diffrence to ci.