+++ e2e/tests/plugins.e2e.js
@@ -35,7 +35,8 @@ describe('plugins', () => {
 
     after(() => cy.eyesClose());
 
-    it('render image plugin toolbar', function() {
+    // eslint-disable-next-line mocha/no-skipped-tests
+    it.skip('render image plugin toolbar', function() {
       cy.log(this);
       cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE).shrinkPlugin();
       cy.eyesCheckWindow(this.test.title);
@@ -90,7 +91,8 @@ describe('plugins', () => {
 
     after(() => cy.eyesClose());
 
-    it('render gallery plugin toolbar', function() {
+    // eslint-disable-next-line mocha/no-skipped-tests
+    it.skip('render gallery plugin toolbar', function() {
       cy.openPluginToolbar(PLUGIN_COMPONENT.GALLERY).shrinkPlugin();
       cy.eyesCheckWindow(this.test.title);
     });
@@ -101,7 +103,9 @@ describe('plugins', () => {
         .get(`[data-hook=${'image-item'}]:first`)
         .get(`[data-hook=${'image-item'}]`)
         .eq(1)
-        .openGalleryAdvancedSettings();
+        .openGalleryAdvancedSettings()
+        .get('.__react_component_tooltip.show')
+        .should('not.exist');
       cy.eyesCheckWindow(this.test.title);
     });
 
@@ -218,7 +222,7 @@ describe('plugins', () => {
       cy.openVideoUploadModal()
         .addVideoFromURI()
         .shrinkPlugin();
-      cy.get(`[data-hook=${PLUGIN_COMPONENT.VIDEO}]:first`);
+      cy.get('[data-loaded=true]');
       cy.eyesCheckWindow(this.test.title);
     });
 
@@ -227,7 +231,7 @@ describe('plugins', () => {
       cy.openVideoUploadModal()
         .addCustomVideo()
         .shrinkPlugin();
-      cy.get(`[data-hook=${PLUGIN_COMPONENT.VIDEO}]:first`);
+      cy.get('[data-loaded=true]');
       cy.eyesCheckWindow(this.test.title);
     });
   });
@@ -250,7 +254,7 @@ describe('plugins', () => {
       cy.openSoundCloudModal()
         .addSoundCloud()
         .shrinkPlugin();
-      cy.get(`[data-hook=${PLUGIN_COMPONENT.SOUND_CLOUD}]:first`);
+      cy.get('[data-loaded=true]');
       cy.eyesCheckWindow(this.test.title);
     });
   });
@@ -317,6 +321,7 @@ describe('plugins', () => {
         .get(`button[data-hook=${PLUGIN_TOOLBAR_BUTTONS.SMALL_CENTER}][tabindex=0]`)
         .click();
       cy.get(`button[data-hook=${PLUGIN_TOOLBAR_BUTTONS.REPLACE}][tabindex=0]`).click();
+      cy.get('.__react_component_tooltip.show').should('not.exist');
       cy.eyesCheckWindow(this.test.title);
     });
   });
@@ -349,6 +354,23 @@ describe('plugins', () => {
 
     it('render file-upload plugin toolbar', function() {
       cy.openPluginToolbar(PLUGIN_COMPONENT.FILE_UPLOAD);
+      cy.eyesCheckWindow(this.test.title);
+    });
+  });
+
+  context('drag and drop', () => {
+    before('load editor', function() {
+      eyesOpen(this);
+      cy.loadEditor('dragAndDrop');
+    });
+
+    after(() => cy.eyesClose());
+
+    it('drag and drop plugins', function() {
+      cy.focusEditor();
+      const src = `[data-hook=${PLUGIN_COMPONENT.IMAGE}] + [data-hook=componentOverlay]`;
+      const dest = `span[data-offset-key="fjkhf-0-0"]`;
+      cy.dragAndDropPlugin(src, dest);
       cy.eyesCheckWindow(this.test.title);
     });
   });