+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -73,7 +73,13 @@ class RichContentEditor extends Component {
       shouldRenderOptimizedImages,
       initialIntent,
       siteDomain,
+      plugins,
     } = this.props;
+
+    const enableDragAndDrop = (plugins || []).some(
+      plugin => plugin.name === 'createBlockDndPlugin'
+    );
+
     this.contextualData = {
       theme,
       t,
@@ -89,6 +95,7 @@ class RichContentEditor extends Component {
       shouldRenderOptimizedImages,
       initialIntent,
       siteDomain,
+      enableDragAndDrop,
     };
   };