+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -41,7 +41,6 @@ class ImageViewer extends React.Component {
 
   getImageUrl(src) {
     const { helpers, shouldRenderOptimizedImages } = this.context || {};
-
     if (!src && helpers?.handleFileSelection) {
       return null;
     }