+++ packages/plugin-sound-cloud/web/src/soundCloud.jsx
@@ -6,6 +6,7 @@ import { mergeStyles, Context } from 'wix-rich-content-common';
 import SoundCloudViewer from './soundCloud-viewer';
 import styles from '../statics/styles/default-sound-cloud-styles.scss';
 import { SOUND_CLOUD_TYPE } from './types';
+import draggableStyle from 'wix-rich-content-common/dist/statics/styles/draggable.scss';
 
 const DEFAULTS = {
   config: {
@@ -82,7 +83,11 @@ class SoundCloud extends Component {
     const { isLoaded } = this.state;
     const overlayText = t('SoundCloudComponent_Overlay');
     return (
-      <div className={classNames(styles.soundCloud_overlay)}>
+      <div
+        className={classNames(styles.soundCloud_overlay, {
+          [draggableStyle.draggable]: this.context.enableDragAndDrop,
+        })}
+      >
         {isLoaded && <span className={styles.soundCloud_overlay_message}>{overlayText}</span>}
       </div>
     );
@@ -90,12 +95,14 @@ class SoundCloud extends Component {
 
   renderPlayer = () => {
     const { componentData } = this.props;
+    const { isPlayable } = this.state;
     return (
       <SoundCloudViewer
         ref={this.setPlayer}
         componentData={componentData}
         onReady={this.handleReady}
         onStart={this.handleStart}
+        isPlayable={isPlayable}
       />
     );
   };
@@ -118,6 +125,7 @@ class SoundCloud extends Component {
         onClick={onClick}
         className={containerClassNames}
         onKeyDown={e => this.onKeyDown(e, onClick)}
+        draggable={this.context.enableDragAndDrop}
       >
         {!isPlayable && this.renderOverlay(this.styles, this.context.t)}
         {this.renderPlayer()}