+++ packages/plugin-video/web/src/video-component.jsx
@@ -6,6 +6,7 @@ import { mergeStyles, Context } from 'wix-rich-content-common';
 import VideoViewer from './video-viewer';
 import styles from '../statics/styles/default-video-styles.scss';
 import { VIDEO_TYPE_LEGACY, VIDEO_TYPE } from './types';
+import draggableStyle from 'wix-rich-content-common/dist/statics/styles/draggable.scss';
 
 const DEFAULTS = {
   config: {
@@ -57,7 +58,11 @@ class VideoComponent extends React.Component {
     const { isLoaded } = this.state;
     const overlayText = t('VideoComponent_Overlay');
     return (
-      <div className={classNames(styles.video_overlay)}>
+      <div
+        className={classNames(styles.video_overlay, {
+          [draggableStyle.draggable]: this.context.enableDragAndDrop,
+        })}
+      >
         {isLoaded && <span className={styles.video_overlay_message}>{overlayText}</span>}
       </div>
     );
@@ -65,12 +70,14 @@ class VideoComponent extends React.Component {
 
   renderPlayer = () => {
     const { componentData, settings } = this.props;
+    const { isPlayable } = this.state;
     return (
       <VideoViewer
         ref={this.setPlayer}
         componentData={componentData}
         settings={settings}
         onReady={this.handleReady}
+        isPlayable={isPlayable}
       />
     );
   };
@@ -93,6 +100,7 @@ class VideoComponent extends React.Component {
         onClick={onClick}
         className={containerClassNames}
         onKeyDown={e => this.onKeyDown(e, onClick)}
+        draggable={this.context.enableDragAndDrop}
       >
         {!isPlayable && this.renderOverlay(this.styles, this.context.t)}
         {this.renderPlayer()}